WCM-832: move campus study course into content modules
