+++ core/docs/changelog/WCM-832.change
@@ -1,1 +1,1 @@
-WCM-832: restore study course block
+WCM-832: move campus study course into content modules
