+++ core/pyproject.toml
@@ -258,7 +258,6 @@ zopeshell = "zeit.cms.cli:zope_shell"
 
 [project.entry-points."fanstatic.libraries"]
 zeit_addcentral = "zeit.addcentral.resources:lib"
-zeit_campus = "zeit.campus.browser.resources:lib"
 zeit_cmp = "zeit.cmp.browser.resources:lib"
 
 zeit_cms = "zeit.cms.browser.resources:lib_css"
@@ -275,6 +274,7 @@ zeit_content_cp = "zeit.content.cp.browser.resources:lib"
 zeit_content_gallery = "zeit.content.gallery.browser.resources:lib"
 zeit_content_image = "zeit.content.image.browser.resources:lib"
 zeit_content_link = "zeit.content.link.browser.resources:lib"
+zeit_content_modules = "zeit.content.modules.browser.resources:lib"
 zeit_content_volume = "zeit.content.volume.browser.resources:lib"
 
 zeit_edit = "zeit.edit.browser.resources:lib_css"
