--- core/src/zeit/campus/__init__.py
@@ -1,1 +1,0 @@
-# python package
