--- core/src/zeit/campus/browser/__init__.py
@@ -1,1 +1,0 @@
-# python package
