--- core/src/zeit/campus/configure.zcml
@@ -1,29 +1,0 @@
-<configure
-  xmlns="http://namespaces.zope.org/zope"
-  xmlns:grok="http://namespaces.zope.org/grok"
-  i18n_domain="zeit.cms">
-
-  <include package="zeit.edit" file="meta.zcml" />
-
-  <grok:grok package="." exclude="browser" />
-
-  <class class=".article.StudyCourse">
-    <require
-      interface=".interfaces.IStudyCourse"
-      permission="zope.View"
-      />
-    <require
-      set_schema=".interfaces.IStudyCourse"
-      permission="zeit.EditContent"
-      />
-  </class>
-
-  <class class=".interfaces.StudyCourse">
-    <!-- Declaring a separate interface feels like overkill -->
-    <require
-      attributes="id title available text href button_text"
-      permission="zope.View"
-      />
-  </class>
-
-</configure>
