+++ core/src/zeit/content/modules/article.py
@@ -1,20 +1,20 @@
 import grokcore.component as grok
 
 from zeit.cms.i18n import MessageFactory as _
-import zeit.campus.interfaces
 import zeit.content.article.edit.block
 import zeit.content.article.edit.interfaces
+import zeit.content.modules.interfaces
 import zeit.edit.block
 
 
-@grok.implementer(zeit.campus.interfaces.IStudyCourse)
+@grok.implementer(zeit.content.modules.interfaces.IStudyCourse)
 class StudyCourse(zeit.edit.block.SimpleElement):
     area = zeit.content.article.edit.interfaces.IEditableBody
     type = 'studycourse'
 
     _course = zeit.cms.content.property.DAVConverterWrapper(
         zeit.cms.content.property.ObjectPathAttributeProperty('.', 'id'),
-        zeit.campus.interfaces.IStudyCourse['course'],
+        zeit.content.modules.interfaces.IStudyCourse['course'],
     )
 
     @property
@@ -22,7 +22,7 @@ class StudyCourse(zeit.edit.block.SimpleElement):
         # Since the values come in via xml config file, we cannot declare the
         # default on the field without major hassle.
         if self._course is None:
-            source = zeit.campus.interfaces.IStudyCourse['course'].source(self)
+            source = zeit.content.modules.interfaces.IStudyCourse['course'].source(self)
             return source.find('default')
         return self._course
