+++ core/src/zeit/content/modules/browser/article.py
@@ -1,11 +1,14 @@
 import zope.formlib.form
 
+import zeit.content.modules.interfaces
 import zeit.edit.browser.form
 
 
 class StudyCourse(zeit.edit.browser.form.InlineForm):
     legend = ''
-    form_fields = zope.formlib.form.FormFields(zeit.campus.interfaces.IStudyCourse).select('course')
+    form_fields = zope.formlib.form.FormFields(zeit.content.modules.interfaces.IStudyCourse).select(
+        'course'
+    )
 
     @property
     def prefix(self):
