+++ core/src/zeit/content/modules/browser/configure.zcml
@@ -3,7 +3,30 @@
   xmlns:browser="http://namespaces.zope.org/browser"
   xmlns:grok="http://namespaces.zope.org/grok"
   i18n_domain="zeit.cms">
-
   <grok:grok package="." />
+
+  <!-- article -->
+  <browser:viewlet
+    for="..interfaces.IStudyCourse"
+    layer="zeit.cms.browser.interfaces.ICMSLayer"
+    view="zope.interface.Interface"
+    manager="zeit.edit.interfaces.IContentViewletManager"
+    name="edit-studycourse"
+    class="zeit.edit.browser.form.FormLoader"
+    permission="zope.View"
+    weight="10"
+    />
+
+  <browser:page
+    for="..interfaces.IStudyCourse"
+    layer="zeit.cms.browser.interfaces.ICMSLayer"
+    name="edit-studycourse"
+    class=".article.StudyCourse"
+    permission="zope.View"
+    />
+
+  <class class="..article.StudyCourse">
+    <implements interface="zeit.edit.interfaces.IFoldable" />
+  </class>
 
 </configure>
