+++ core/src/zeit/content/modules/browser/resources/editor.css
