+++ core/src/zeit/content/modules/configure.zcml
@@ -29,4 +29,23 @@
       />
   </class>
 
+  <class class=".article.StudyCourse">
+    <require
+      interface=".interfaces.IStudyCourse"
+      permission="zope.View"
+      />
+    <require
+      set_schema=".interfaces.IStudyCourse"
+      permission="zeit.EditContent"
+      />
+  </class>
+
+  <class class=".interfaces.StudyCourse">
+    <!-- Declaring a separate interface feels like overkill -->
+    <require
+      attributes="id title available text href button_text"
+      permission="zope.View"
+      />
+  </class>
+
 </configure>
